Lots of cosmetic stuff!
Updated KeyBuffer to have a nice cursor.
AppStateGameMode can now get the contents of the KeyBuffer to process
(although nothing is done with it yet).
Made inventory windows generally nicer looking, but there's still a lot
of work to do there.
Changed any TODO instance to a Doxygen @todo instead to make it easier
to track them.
While doing that, cleaned up quite a few TODOs that were done.