Actions cleanup and refactoring
Changed Entity actions so that the Thing::do_action_* calls use a
double-dispatch which forces the main Thing class behavior to run first
before subclass behavior is executed.  (This allows to ensure certain
overall rules, such as that a magically locked item cannot be
unwielded.)

Renamed most of the Thing::can_be_* methods to better match what they
really mean (drinkable_by, edible_by, etc.).  These methods are supposed
to refer to generic properties, such as that a potion can be drank by a
human, not that a *particular* potion can be drank by this *particular*
human, and so they should probably become class statics down the line
instead of virtuals... but that's something to tackle later.

Some other renaming here and there for clarification.