ARGH.  Look BEFORE you leap.
Reminder to self: check to see that code does NOT crash with an
exception before committing!  Having said that... this code will still
crash with an exception.  I'm getting a std::out_of_range that I've
unsuccessfully tried to fix four times now, and I'm late getting home,
so I'm stopping here until I can figure out what the hell is wrong.

Hey, guess what you can't use as the key for a std::map?  An enum class!
Changed the AttributeSet code to cast to and from unsigned ints instead.

Also added the ability to wield items, but this is COMPLETELY TOTALLY
UNTESTED so far.