Lighting refactor, small map generation change
Refactored lighting to get rid of a dynamic_cast.
Added a random chance of a MapCorridor connecting with a previous map
feature if it happens to end right there; this allows for some loops in
the map (where previously it took the form of an N-ary tree with no
cycles in it at all).