Splitting up some graphics
I realized that if I really want to be able to add new stuff on the fly,
I need to split up the graphics into separate PNG files. So that has
been done.

I have also started work on the MapTileMetadata class, which will
replace the old MapTileType enum class.