Ton of changes
Deleted AISessile and AIStrategy -- they will not be used now that Lua
scripting will be done.

Added some functions to do Lua tracing. They are not complete.

Added the NEW macro which is used in the Visual C++ leak tracker, and
changed all calls to "new" to use it instead.

Did a bunch of restructuring in an attempt to take care of memory leaks
that did not actually exist due to my running the leak checker wrong.
But the restructuring isn't BAD per se so I'll leave it in there. Might
even be faster in the end.