New Pimpl helpers
Templated Pimpl classes get rid of a lot of boilerplate -- but it also
means moving the Impl for classes that use it into their own files,
because the template doesn't work with forward declaration. Still, I
think it's worth it in the end.

I have started with the Action class, and will refactor other code as
time goes on.