Created PropertyDictionary
The PropertyDictionary is a wrapper class for an unordered_map of
std::string to boost::any. The class makes sure that we only get/set the
particular data types we want in the map.

This makes ConfigSettings super trivial -- it now just inherits from
PropertyDictionary. Likewise, Metadata will be changed to get
intrinsic/default properties and store them in PropertyDictionary
objects, which will eliminate doubled code that handles that in
ThingMetadata and MapTileMetadata separately right now.