An attempt at a small change results in a gigantic refactor.
I had this simple idea to keep track of the children that are derived
from a particular parent (thing, maptile, etc). But because this
required different Metadata instances to access one another, it meant
refactoring how Metadata was stored in the code. And one thing led to
another, and eventually there were changes all over the codebase to
accommodate this. And the worst part? I haven't even implemented the
original "child list" feature that started this whole thing!