Lighting optimization
Lighting is no longer calculated in the weird "two-pass" way it was
being done before. As a result is it much faster to calculate, and will
allow for nifty effects down the line like semi-opaque objects.

Also fixed a gradient quad bug that was causing subtle issues with wall
lighting.

Wall lighting is a little strange right now, probably because I'm
fudging it a bit. Will have to investigate later.