Big refactor of Actions to make them easier to register in a central handler
In the end, Actions will have to be callable by Lua scripts, and we want
it to be (relatively) easy to add new Actions as necessary. Thus, the
Action interface has been refactored a bit to make it easier to
implement a factory method and class registration down the line.

Action states (prebegin, begin, finish, abort) have also been split into
non-virtual and virtual portions, to make it easier to pull out code
that is shared across multiple actions (such as verifying that the
target is within reach).

A CRTP class was added for a static class create() function; this will
be registered into a map of some sort eventually for easy Action
creation via a factory method.