Fixed a problem with Lua class inheritance, and fixed MAJOR TileSheet bugs
A problem with Lua class inheritance meant that the class parent chain
was not being properly registered. When that was fixed, it unveiled a
ton of bugs with the logic used to place tiles on the TileSheet (because
suddenly the algorithm was picking up tiles that it wasn't before, and
that caused the total number of tiles to exceed 32).

Might still be more logic bugs, and there are definitely a lot of trace
prints I need to remove, but for now this is looking good.