Enormous changes -- got rid of wchar_t, added 64-bit support, incorporated boost::locale
The 64-bit builds are panicking somewhere in a LuaJIT call, but at least
everything builds properly, and the application starts. Looks like it
has to do with lua_Integer being declared as 64 bits instead of 32 bits.
Also added required DLLs for different kinds of builds in the "dll"
directory.
Introduced bug: Everything In The Message Log Is In Title Case Now. Will
fix later.