Bunch more event handling work
Re-architected the whole GUI object event system so it isn't
hierarchical any more. Was just too much of a hassle to get working
right, and too complicated.

New event system is not quite working right, but it's a start.

Changed _NVI suffixes to _V: Easier to read, faster to type, and makes a
bit more sense ("_NVI" implied that it was the Non-Virtual part of the
NVI pattern, rather than the virtual part)