Renaming and event handling updates
Renamed `onEvent` to `onEvent_NV`, `onEvent_V` to `onEvent`, and
`onEvent_V2` to `onEvent_V`. This improves code legibility overall.
Updated GUIObject so it registers/deregisters from App events when added
to / removed from a parent.
Added the ability for an object to see if a point is inside it but *not*
inside its children (so it knows whether to process a mouse event it
sees).
Changed some snake_case stuff to camelCase.