Replaced "Mu" with "Void"
"Mu" was cute and all, but cute != readable. "Void" is a much better
description of its function within the engine.

Also changed the way to access the Void EntityId -- it's a static object
instead of a static method. I have no idea why it was a static method in
the first place -- there was no reason for it I can see, and it slowed
things down (in the debug build at least).